refactor: aligns import of user
